Refactor cloudinary api route structure
