+++ pages/api/cloudinary/browse.ts
@@ -1,0 +1,15 @@
+import { NextApiRequest, NextApiResponse } from 'next';
+import { v2 as cloudinary } from 'cloudinary';
+
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
+
+const browseImagesHandler = async (
+  req: NextApiRequest,
+  res: NextApiResponse
+) => {};
+
+export default browseImagesHandler;
