+++ pages/blog/create.tsx
@@ -4,7 +4,7 @@ import Router from 'next/router';
 
 import { useSession } from 'next-auth/react';
 import { useFetchStatus } from '@/hooks/useLoadingIndicator';
-import { uploadImage } from '@/lib/blog';
+import { uploadImage } from '@/lib/cloudinary';
 
 import LoadingTriangle from '@/components/LoadingTriangle';
 import Container from '@/components/Container';
